+++ cmd/hlk/client/hlk-cli/errors.go
@@ -1,0 +1,19 @@
+package main
+
+const (
+	errPrefix = "cmd/hlk/client/hlk-cli = "
+)
+
+type SError struct {
+	str string
+}
+
+func (err *SError) Error() string {
+	return errPrefix + err.str
+}
+
+var (
+	ErrMkdirPath     = &SError{"mkdir path"}
+	ErrRetryNum      = &SError{"retry num"}
+	ErrUnknownAction = &SError{"unknown action"}
+)